remove relayer traits from fuel-core-interfaces
