+++ crates/fuel-core-interfaces/src/lib.rs
@@ -3,6 +3,5 @@
 pub mod bft;
 pub mod block_importer;
 pub mod poa_coordinator;
-pub mod relayer;
 pub mod signer;
 pub mod txpool;
