+++ crates/services/relayer/src/lib.rs
@@ -12,6 +12,7 @@ mod relayer;
 
 #[cfg(any(test, feature = "test-helpers"))]
 pub mod mock_db;
+pub mod ports;
 #[cfg(any(test, feature = "test-helpers"))]
 pub mod test_helpers;
